+++ mission/chapter10/--template/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "axios": "^1.7.9",
     "react": "^19.2.0",
-    "react-dom": "^19.2.0"
+    "react-dom": "^19.2.0",
+    "react-router-dom": "^7.9.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.39.1",
